replace: more tests.
Some of the replace results for zero-width matches are a bit weird,
but they match what Perl and Ruby do. For example, it seems like it
would, imo, be much more intuitive if we had the following:

    replace("abcd", r"b?", "^") == "^a^c^d^"
    replace("abcd", r"b?c?", "^") == "^a^d^"
    replace("abcd", r"[bc]?", "^") == "^a^^d^"

However, I'm hesitant to break with the behavior of Perl and Ruby.
I wonder to what extent that behavior is simply based on being a
bit simpler to implement.